Pop up message for delete client

+++ frontend/src/components/updateClient.vue
@@ -48,7 +48,7 @@ export default {
     axios
       .get(
         import.meta.env.VITE_ROOT_API +
-          `/primarydata/id/${this.$route.params.id}`
+        `/primarydata/id/${this.$route.params.id}`
       )
       .then((resp) => {
         let data = resp.data[0];
@@ -69,7 +69,7 @@ export default {
     axios
       .get(
         import.meta.env.VITE_ROOT_API +
-          `/eventdata/client/${this.$route.params.id}`
+        `/eventdata/client/${this.$route.params.id}`
       )
       .then((resp) => {
         let data = resp.data;
@@ -113,7 +113,7 @@ export default {
           axios
             .get(
               import.meta.env.VITE_ROOT_API +
-                `/eventdata/client/${this.$route.params.id}`
+              `/eventdata/client/${this.$route.params.id}`
             )
             .then((resp) => {
               let data = resp.data;
@@ -127,13 +127,16 @@ export default {
       });
     },
     deleteUser() {
-      let apiURL = import.meta.env.VITE_ROOT_API + `/eventdata/removeclient/${this.id}`;
-      axios.delete(apiURL).then(() => {
-      axios.delete(import.meta.env.VITE_ROOT_API + `/primarydata/delete/${this.id}`).then(() => {
-        alert("Client has been deleted.");
-        this.$router.back()
+      this.$confirm("Are you sure?").then(() => {
+        let apiURL = import.meta.env.VITE_ROOT_API + `/eventdata/removeclient/${this.id}`;
+        axios.delete(apiURL).then(() => {
+          axios.delete(import.meta.env.VITE_ROOT_API + `/primarydata/delete/${this.id}`).then(() => {
+            alert("Client has been deleted.");
+            this.$router.back()
+          });
+        });
+
       });
-    });
     },
   },
   validations() {
@@ -166,18 +169,12 @@ export default {
             <label class="block">
               <span class="text-gray-700">First Name</span>
               <span style="color:#ff0000">*</span>
-              <input
-                type="text"
+              <input type="text"
                 class="w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-300 focus:ring focus:ring-indigo-200 focus:ring-opacity-50"
-                placeholder
-                v-model="client.firstName"
-              />
+                placeholder v-model="client.firstName" />
               <span class="text-black" v-if="v$.client.firstName.$error">
-                <p
-                  class="text-red-700"
-                  v-for="error of v$.client.firstName.$errors"
-                  :key="error.$uid"
-                >{{ error.$message }}!</p>
+                <p class="text-red-700" v-for="error of v$.client.firstName.$errors" :key="error.$uid">{{ error.$message
+                }}!</p>
               </span>
             </label>
           </div>
@@ -186,12 +183,9 @@ export default {
           <div class="flex flex-col">
             <label class="block">
               <span class="text-gray-700">Middle Name</span>
-              <input
-                type="text"
+              <input type="text"
                 class="w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-300 focus:ring focus:ring-indigo-200 focus:ring-opacity-50"
-                placeholder
-                v-model="client.middleName"
-              />
+                placeholder v-model="client.middleName" />
             </label>
           </div>
 
@@ -200,18 +194,12 @@ export default {
             <label class="block">
               <span class="text-gray-700">Last Name</span>
               <span style="color:#ff0000">*</span>
-              <input
-                type="text"
+              <input type="text"
                 class="w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-300 focus:ring focus:ring-indigo-200 focus:ring-opacity-50"
-                placeholder
-                v-model="client.lastName"
-              />
+                placeholder v-model="client.lastName" />
               <span class="text-black" v-if="v$.client.lastName.$error">
-                <p
-                  class="text-red-700"
-                  v-for="error of v$.client.lastName.$errors"
-                  :key="error.$uid"
-                >{{ error.$message }}!</p>
+                <p class="text-red-700" v-for="error of v$.client.lastName.$errors" :key="error.$uid">{{ error.$message
+                }}!</p>
               </span>
             </label>
           </div>
@@ -220,18 +208,12 @@ export default {
           <div class="flex flex-col">
             <label class="block">
               <span class="text-gray-700">Email</span>
-              <input
-                type="email"
+              <input type="email"
                 class="w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-300 focus:ring focus:ring-indigo-200 focus:ring-opacity-50"
-                pattern="[a-z0-9._%+-]+@[a-z0-9.-]+\.[a-z]{2,}$"
-                v-model="client.email"
-              />
+                pattern="[a-z0-9._%+-]+@[a-z0-9.-]+\.[a-z]{2,}$" v-model="client.email" />
               <span class="text-black" v-if="v$.client.email.$error">
-                <p
-                  class="text-red-700"
-                  v-for="error of v$.client.email.$errors"
-                  :key="error.$uid"
-                >{{ error.$message }}!</p>
+                <p class="text-red-700" v-for="error of v$.client.email.$errors" :key="error.$uid">{{ error.$message }}!
+                </p>
               </span>
             </label>
           </div>
@@ -240,18 +222,12 @@ export default {
             <label class="block">
               <span class="text-gray-700">Phone Number</span>
               <span style="color:#ff0000">*</span>
-              <input
-                type="text"
+              <input type="text"
                 class="w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-300 focus:ring focus:ring-indigo-200 focus:ring-opacity-50"
-                pattern="[0-9]{3}[0-9]{3}[0-9]{4}"
-                v-model="client.phoneNumbers[0].primaryPhone"
-              />
+                pattern="[0-9]{3}[0-9]{3}[0-9]{4}" v-model="client.phoneNumbers[0].primaryPhone" />
               <span class="text-black" v-if="v$.client.phoneNumbers[0].primaryPhone.$error">
-                <p
-                  class="text-red-700"
-                  v-for="error of v$.client.phoneNumbers[0].primaryPhone.$errors"
-                  :key="error.$uid"
-                >{{ error.$message }}!</p>
+                <p class="text-red-700" v-for="error of v$.client.phoneNumbers[0].primaryPhone.$errors"
+                  :key="error.$uid">{{ error.$message }}!</p>
               </span>
             </label>
           </div>
@@ -259,12 +235,9 @@ export default {
           <div class="flex flex-col">
             <label class="block">
               <span class="text-gray-700">Alternative Phone Number</span>
-              <input
-                type="text"
+              <input type="text"
                 class="w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-300 focus:ring focus:ring-indigo-200 focus:ring-opacity-50"
-                pattern="[0-9]{3}[0-9]{3}[0-9]{4}"
-                v-model="client.phoneNumbers[0].secondaryPhone"
-              />
+                pattern="[0-9]{3}[0-9]{3}[0-9]{4}" v-model="client.phoneNumbers[0].secondaryPhone" />
             </label>
           </div>
         </div>
@@ -276,22 +249,18 @@ export default {
           <div class="flex flex-col">
             <label class="block">
               <span class="text-gray-700">Address Line 1</span>
-              <input
-                type="text"
+              <input type="text"
                 class="w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-300 focus:ring focus:ring-indigo-200 focus:ring-opacity-50"
-                v-model="client.address.line1"
-              />
+                v-model="client.address.line1" />
             </label>
           </div>
           <!-- form field -->
           <div class="flex flex-col">
             <label class="block">
               <span class="text-gray-700">Address Line 2</span>
-              <input
-                type="text"
+              <input type="text"
                 class="w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-300 focus:ring focus:ring-indigo-200 focus:ring-opacity-50"
-                v-model="client.address.line2"
-              />
+                v-model="client.address.line2" />
             </label>
           </div>
           <!-- form field -->
@@ -299,11 +268,9 @@ export default {
             <label class="block">
               <span class="text-gray-700">City</span>
               <span style="color:#ff0000">*</span>
-              <input
-                type="text"
+              <input type="text"
                 class="w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-300 focus:ring focus:ring-indigo-200 focus:ring-opacity-50"
-                v-model="client.address.city"
-              />
+                v-model="client.address.city" />
             </label>
           </div>
           <div></div>
@@ -311,22 +278,18 @@ export default {
           <div class="flex flex-col">
             <label class="block">
               <span class="text-gray-700">County</span>
-              <input
-                type="text"
+              <input type="text"
                 class="w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-300 focus:ring focus:ring-indigo-200 focus:ring-opacity-50"
-                v-model="client.address.county"
-              />
+                v-model="client.address.county" />
             </label>
           </div>
           <!-- form field -->
           <div class="flex flex-col">
             <label class="block">
               <span class="text-gray-700">Zip Code</span>
-              <input
-                type="text"
+              <input type="text"
                 class="w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-300 focus:ring focus:ring-indigo-200 focus:ring-opacity-50"
-                v-model="client.address.zip"
-              />
+                v-model="client.address.zip" />
             </label>
           </div>
           <div></div>
@@ -335,25 +298,15 @@ export default {
         <!-- grid container -->
         <div class="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-4 gap-x-6 gap-y-10">
           <div class="flex justify-between mt-10 mr-20">
-            <button
-              @click="handleClientUpdate"
-              type="submit"
-              class="bg-red-700 text-white rounded"
-            >Update Client</button>
+            <button @click="handleClientUpdate" type="submit" class="bg-red-700 text-white rounded">Update
+              Client</button>
           </div>
           <div class="flex justify-between mt-10 mr-20">
-            <button
-              type="reset"
-              class="border border-red-700 bg-white text-red-700 rounded"
-              @click="$router.go(-1)"
-            >Go back</button>
+            <button type="reset" class="border border-red-700 bg-white text-red-700 rounded" @click="$router.go(-1)">Go
+              back</button>
           </div>
           <div class="flex justify-between mt-10 mr-20">
-            <button
-              type="submit"
-              class="bg-red-700 text-white rounded"
-              @click="deleteUser"
-            >Delete Client</button>
+            <button type="submit" class="bg-red-700 text-white rounded" @click="deleteUser">Delete Client</button>
           </div>
         </div>
 
@@ -384,17 +337,10 @@ export default {
             <label class="typo__label">Select Events to be added</label>
             <VueMultiselect
               class="w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-300 focus:ring focus:ring-indigo-200 focus:ring-opacity-50"
-              v-model="eventsChosen"
-              :options="eventData"
-              :multiple="true"
-              label="eventName"
-            ></VueMultiselect>
+              v-model="eventsChosen" :options="eventData" :multiple="true" label="eventName"></VueMultiselect>
             <div class="flex justify-between">
-              <button
-                @click="addToEvent"
-                type="submit"
-                class="mt-5 bg-red-700 text-white rounded"
-              >Add Client to Events</button>
+              <button @click="addToEvent" type="submit" class="mt-5 bg-red-700 text-white rounded">Add Client to
+                Events</button>
             </div>
           </div>
         </div>
@@ -402,4 +348,6 @@ export default {
     </div>
   </main>
 </template>
-<style src="vue-multiselect/dist/vue-multiselect.css"></style>
+<style src="vue-multiselect/dist/vue-multiselect.css">
+
+</style>
